+++ crates/nix_rs/src/info.rs
@@ -1,5 +1,6 @@
 //! Information about the user's Nix installation
 use serde::{Deserialize, Serialize};
+use tokio::sync::OnceCell;
 
 use crate::{config::NixConfig, env::NixEnv, version::NixVersion};
 
@@ -12,11 +13,23 @@ pub struct NixInfo {
     pub nix_env: NixEnv,
 }
 
+static NIX_INFO: OnceCell<Result<NixInfo, NixInfoError>> = OnceCell::const_new();
+
 impl NixInfo {
+    /// Get the once version  of `NixInfo`
+    pub async fn get() -> &'static Result<NixInfo, NixInfoError> {
+        NIX_INFO
+            .get_or_init(|| async {
+                let nix_config = NixConfig::get().await.as_ref()?;
+                let info = NixInfo::new(nix_config.clone()).await?;
+                Ok(info)
+            })
+            .await
+    }
+
     /// Determine [NixInfo] on the user's system
-    pub async fn from_nix(nix_cmd: &crate::command::NixCmd) -> Result<NixInfo, NixInfoError> {
-        let nix_version = NixVersion::from_nix(nix_cmd).await?;
-        let nix_config = NixConfig::from_nix(nix_cmd).await?;
+    pub async fn new(nix_config: NixConfig) -> Result<NixInfo, NixInfoError> {
+        let nix_version = NixVersion::from_nix().await?;
         let nix_env = NixEnv::detect().await?;
         Ok(NixInfo {
             nix_version,
@@ -30,6 +43,9 @@ impl NixInfo {
 pub enum NixInfoError {
     #[error("Nix command error: {0}")]
     NixCmdError(#[from] crate::command::NixCmdError),
+
+    #[error("Nix command error: {0}")]
+    NixCmdErrorStatic(#[from] &'static crate::command::NixCmdError),
 
     #[error("Nix environment error: {0}")]
     NixEnvError(#[from] crate::env::NixEnvError),