+++ crates/omnix-gui/src/app/state/mod.rs
@@ -9,6 +9,7 @@ use dioxus::prelude::*;
 use dioxus_signals::{Readable, Signal, Writable};
 use nix_health::NixHealth;
 use nix_rs::{
+    config::NixConfig,
     flake::{url::FlakeUrl, Flake},
     info::NixInfo,
 };
@@ -116,13 +117,17 @@ impl AppState {
             let mut flake = self.flake;
             let mut flake_cache = self.flake_cache;
             let _ = use_resource(move || async move {
+                let nixcmd = nix_rs::command::NixCmd::get().await;
                 let flake_url = flake_url.read().clone();
                 let refresh = *flake_refresh.read();
                 if let Some(flake_url) = flake_url {
                     let flake_url_2 = flake_url.clone();
                     tracing::info!("Updating flake [{}] refresh={} ...", &flake_url, refresh);
                     let res = Datum::refresh_with(&mut flake, async move {
-                        Flake::from_nix(&nix_rs::command::NixCmd::default(), flake_url_2)
+                        let nix_config = NixConfig::from_nix(nixcmd)
+                            .await
+                            .map_err(|e| Into::<SystemError>::into(e.to_string()))?;
+                        Flake::from_nix(nixcmd, &nix_config, flake_url_2)
                             .await
                             .map_err(|e| Into::<SystemError>::into(e.to_string()))
                     })
@@ -167,11 +172,13 @@ impl AppState {
                 let refresh = *nix_info_refresh.read();
                 tracing::info!("Updating nix info [{}] ...", refresh);
                 Datum::refresh_with(&mut nix_info, async {
-                    NixInfo::from_nix(&nix_rs::command::NixCmd::default())
+                    let cfg = NixConfig::get()
                         .await
-                        .map_err(|e| SystemError {
-                            message: format!("Error getting nix info: {:?}", e),
-                        })
+                        .as_ref()
+                        .map_err(|e| Into::<SystemError>::into(e.to_string()))?;
+                    NixInfo::new(cfg.clone()).await.map_err(|e| SystemError {
+                        message: format!("Error getting nix info: {:?}", e),
+                    })
                 })
                 .await;
             });